- fixed multilanguage config issue for not 100% correct working language switchers. Fallback is now correctly added - small fix in mediahandler, it actually never appeared as bug, but is cleaner now - fix in migrator to reflect the new order prices - a little more compatible search in orders - added allowed chars to keywords if you search for products (# and /) - fixed unexpected T_PAAMAYIM_NEKUDOTAYIM error for people using php5.2 (yeh we have an heart for you guys) - fixed that the revenue update button reappeared for orders with total=0.0 - added to the config default the normal settings for the virtuemart frontpage, so the config dont need to be stored to see it directly after installation - fix for virtuemart_Category_id= 0 is not published - added fix for title (thx Pro) - removed some vmdebugs
git-svn-id: https://dev.virtuemart.net/svn/virtuemart/branches/com_virtuemart2.0.14mp@7071 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ administrator/components/com_virtuemart/models/updatesmigration.php
@@ -258,9 +258,7 @@ class VirtueMartModelUpdatesMigration extends JModel {
 				$url = '/plugins/vmshipment';
 			}
 
-
 			if (!class_exists ('plgVmShipmentWeight_countries')) require(JPATH_ROOT . DS . $url . DS . 'weight_countries.php');
-			$this->_tablename = '#__virtuemart_' . $this->_psType . '_plg_' . $this->_name;
 			$this->installPluginTable('plgVmShipmentWeight_countries','#__virtuemart_shipment_plg_weight_countries','Shipment Weight Countries Table');
 		}
 
@@ -300,8 +298,10 @@ class VirtueMartModelUpdatesMigration extends JModel {
 				$query .= '`' . $fieldname . '` ' . $fieldtype . " , ";
 			}
 		} else {
-			$SQLfields = $className::getTableSQLFields ();
-			$loggablefields = $className::getTableSQLLoggablefields ();
+			$SQLfields = call_user_func($className."::getTableSQLFields");
+			//$SQLfields = $className::getTableSQLFields ();
+		//	$loggablefields = $className::getTableSQLLoggablefields ();
+			$loggablefields = call_user_func($className."::getTableSQLLoggablefields");
 			foreach ($SQLfields as $fieldname => $fieldtype) {
 				$query .= '`' . $fieldname . '` ' . $fieldtype . " , ";
 			}
@@ -536,14 +536,21 @@ class VirtueMartModelUpdatesMigration extends JModel {
 	private function deleteMediaThumbFolder($type,$resized='resized'){
 
 		if(!empty($resized)) $resized = DS.$resized;
-		$path = JPATH_ROOT.DS.str_replace('/',DS,VmConfig::get($type)).$resized;
-		$msg = JFolder::delete($path);
-		if(!$msg){
-			vmWarn('Problem deleting '.$type);
+		$typePath = VmConfig::get($type);
+		if(!empty($typePath)){
+			$path = JPATH_ROOT.DS.str_replace('/',DS,$typePath).$resized;
+			$msg = JFolder::delete($path);
+			if(!$msg){
+				vmWarn('Problem deleting '.$type);
+			}
+			if(!class_exists('JFile')) require(JPATH_VM_LIBRARIES.DS.'joomla'.DS.'filesystem'.DS.'file.php');
+			$msg = JFolder::create($path);
+			return $msg;
+		} else {
+
+			return 'Config path for '.$type.' empty';
 		}
-		if(!class_exists('JFile')) require(JPATH_VM_LIBRARIES.DS.'joomla'.DS.'filesystem'.DS.'file.php');
-		$msg = JFolder::create($path);
-		return $msg;
+
 	}
 
 }
